Refactor tokmd-node binding internals
